moved button components to the directory /components/f-styled/buttons to match Axel's PR
